+++ test.d
@@ -8,6 +8,7 @@ import std.algorithm;
 import std.string;
 import std.exception;
 
+
 int main ()
 {
     return TestRunner().run;
@@ -26,9 +27,15 @@ struct TestRunner
 
         foreach (const clang ; matrix.clangs)
         {
+            import std.string;
+
             activate(clang);
-            writeln("Testing with libclang version ", clang.version_);
+
+            auto output = execute(["./bin/dstep", "--clang-version"]);
+
+            writeln("Testing with ", strip(output.output));
             result += unitTest();
+            stdout.flush();
         }
 
         return result;
@@ -71,7 +78,10 @@ struct TestRunner
     {
         writeln("Running unit tests ");
 
-        auto result = executeShell("dub test");
+        version (Win64)
+            auto result = executeShell("dub test --arch=x86_64");
+        else
+            auto result = executeShell("dub test");
 
         if (result.status != 0)
             writeln(result.output);
@@ -81,7 +91,10 @@ struct TestRunner
 
     void build ()
     {
-        auto result = executeShell("dub build");
+        version (Win64)
+            auto result = executeShell("dub build --arch=x86_64");
+        else
+            auto result = executeShell("dub build");
 
         if (result.status != 0)
         {
@@ -112,7 +125,7 @@ struct Clang
     else version (Windows)
     {
         enum extension = ".dll";
-        enum prefix = "";
+        enum prefix = "lib";
     }
 
     else version (FreeBSD)
@@ -162,6 +175,7 @@ struct ClangMatrix
 
             writeln("Downloading clang ", clang.version_);
             mkdirRecurse(basePath);
+            stdout.flush();
             download(clang);
         }
     }
@@ -176,6 +190,7 @@ struct ClangMatrix
             writeln("Extracting clang ", clang.version_);
             extractArchive(clang);
             extractLibclang(clang);
+            stdout.flush();
             clean();
         }
     }
@@ -206,7 +221,10 @@ private:
         auto dest = clangPath();
         mkdirRecurse(dest);
 
-        auto result = execute(["tar", "--strip-components=1", "-C", dest, "-xf", src]);
+        version (Posix)
+            auto result = execute(["tar", "--strip-components=1", "-C", dest, "-xf", src]);
+        else
+            auto result = execute(["7z", "x", src, "-y", format("-o%s", dest)]);
 
         if (result.status != 0)
             throw new ProcessException("Failed to extract archive");
@@ -227,7 +245,11 @@ private:
 
     void extractLibclang (const ref Clang clang)
     {
-        auto src = buildNormalizedPath(clangPath, "lib", clang.libclang);
+        version (Posix)
+            auto src = buildNormalizedPath(clangPath, "lib", clang.libclang);
+        else
+            auto src = buildNormalizedPath(clangPath, "bin", clang.libclang);
+
         auto dest = buildNormalizedPath(workingDirectory, clang.versionedLibclang);
 
         copy(src, dest);
@@ -359,12 +381,27 @@ private:
                 static assert(false, "Only 64bit versions of OS X are supported");
         }
 
-        else version (Windows)
+        else version (Win32)
         {
             return [
+                // Clang("3.8.1", "http://llvm.org/releases/3.8.1/", "LLVM-3.8.1-win32.exe"),
+                Clang("3.8.0", "http://llvm.org/releases/3.8.0/", "LLVM-3.8.0-win32.exe"),
+                // Clang("3.7.1", "http://llvm.org/releases/3.7.1/", "LLVM-3.7.1-win32.exe"),
+                Clang("3.7.0", "http://llvm.org/releases/3.7.0/", "LLVM-3.7.0-win32.exe"),
+                // Clang("3.6.2", "http://llvm.org/releases/3.6.2/", "LLVM-3.6.2-win32.exe"),
+                // Clang("3.6.1", "http://llvm.org/releases/3.6.1/", "LLVM-3.6.1-win32.exe"),
+                Clang("3.6.0", "http://llvm.org/releases/3.6.0/", "LLVM-3.6.0-win32.exe"),
                 Clang("3.5.0", "http://llvm.org/releases/3.5.0/", "LLVM-3.5.0-win32.exe"),
-                Clang("3.4.1", "http://llvm.org/releases/3.4.1/", "LLVM-3.4.1-win32.exe"),
+                // Clang("3.4.1", "http://llvm.org/releases/3.4.1/", "LLVM-3.4.1-win32.exe"),
                 Clang("3.4", "http://llvm.org/releases/3.4/", "LLVM-3.4-win32.exe")
+            ];
+        }
+
+        else version (Win64)
+        {
+            return [
+                Clang("3.8.0", "http://llvm.org/releases/3.8.0/", "LLVM-3.8.0-win64.exe"),
+                Clang("3.7.0", "http://llvm.org/releases/3.7.0/", "LLVM-3.7.0-win64.exe"),
             ];
         }